// ui/src/types/generated/Block.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BlockKind } from "./BlockKind";
import type { BlockPayload } from "./BlockPayload";
import type { ByteRange } from "./ByteRange";

export type Block = { kind: BlockKind, byte_range: ByteRange, payload: BlockPayload, };