// ui/src/types/generated/BlockError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BlockId } from "./BlockId";

export type BlockError = { "Overlapping": BlockId } | { "GapBefore": BlockId } | { "InvalidYaml": string } | { "DuplicateId": BlockId } | { "MissingRequiredField": string };