// ui/src/types/generated/CommentAnchor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CommentAnchor = { kind: "html_selection", start_offset: number, end_offset: number, snapshot_text: string, } | { kind?: "text_selection", block_id: string | null, start_offset: number, end_offset: number, };