// ui/src/types/generated/BlockKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BlockKind = "Frontmatter" | "Heading" | "Paragraph" | "List" | "BlockQuote" | "CodeBlock" | "HtmlBlock" | "Table" | "FootnoteDefinition" | "LinkRefDefinition" | "ThematicBreak" | "VellumLiveQuery" | "VellumResult";