// ui/src/types/generated/BlockPatch.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BlockEdit } from "./BlockEdit";
import type { BlockError } from "./BlockError";
import type { BlockId } from "./BlockId";
import type { BlockKind } from "./BlockKind";
import type { ByteRange } from "./ByteRange";

export type BlockPatch = { block_id: BlockId, parsed_kind: BlockKind, original_byte_range: ByteRange | null, edit: BlockEdit, dirty: boolean, error: BlockError | null, };