// ui/src/types/generated/IdentityMap.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BlockIdentity } from "./BlockIdentity";

export type IdentityMap = { source_hash: [number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number, number], block_ids: Array<BlockIdentity>, };