// ui/src/types/generated/Inline.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Inline = { "Text": string } | { "Strong": Array<Inline> } | { "Emphasis": Array<Inline> } | { "Code": string } | { "Link": { href: string, title: string | null, body: Array<Inline>, } } | { "Image": { src: string, title: string | null, alt: string, } } | "HardBreak" | "SoftBreak" | { "Html": string };